refactor(api): split api.ts (1517 LOC) into per-resource modules [Phase3 B1]
Pure file-move refactor. src/lib/api.ts becomes a 9-line barrel
(`export * from './api/<resource>'`); the implementation moves into
src/lib/api/{core,transactions,documents,ingest,reports,merchants,places,
products,brands}.ts. Shared client/utils/types/mappers live in core; every
resource module imports only from ./core (no cycles). Because `../lib/api`
resolves to the barrel, all 17 importers are untouched.

Export-surface verified: zero exports dropped; the only additions are 5
core utilities (client, unwrap, etagFrom, genIdempotencyKey, compressImage)
that were module-private in the monolith and must cross the new module
boundary. (Mildly widens the barrel surface; no caller uses them.)

Behavior-neutral. tsc passing across all 17 importers is the proof (no
runtime logic changed). Verified in-browser (frontend-test-runner):
dashboard / ledger / receipt / brand fetch paths all load, no broken
imports, no failed API requests. Build+tsc+lint(0).
